backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
opencv-python==4.8.0.76
face-recognition==1.3.0
numpy==1.25.2
python-dotenv==1.0.0